backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.2
numpy<2.0
python-multipart==0.0.9
sentence-transformers==2.7.0
chromadb==0.5.0
transformers==4.41.2
torch==2.3.0
peft==0.11.1
accelerate==0.30.1
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
